requirements.txt
asgiref==3.7.2
cfgv==3.4.0
distlib==0.3.8
Django==4.2
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.14.0
dnspython==2.4.2
filelock==3.13.1
identify==2.5.33
iniconfig==2.0.0
mongoadmin==0.2
mongodbforms==0.3
mongoengine==0.27.0
mongomock==4.1.2
nodeenv==1.8.0
packaging==23.2
platformdirs==4.1.0
pluggy==1.3.0
pre-commit==3.6.0
pymongo==4.6.1
pytest==7.4.4
pytest-django==4.7.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
ruff==0.1.9
sentinels==1.0.0
sqlparse==0.4.4
virtualenv==20.25.0
